Moves ShootProjectile to shared
